+++ Libplanet.Tests/Blockchain/Renderers/DelayedRendererTest.cs
@@ -29,7 +29,7 @@ namespace Libplanet.Tests.Blockchain.Renderers
             HashAlgorithmType hashAlgorithm = HashAlgorithmType.Of<SHA256>();
             var chainA = new Block<DumbAction>[10];
             var chainB = new Block<DumbAction>[chainA.Length];
-            chainA[0] = chainB[0] = TestUtils.MineGenesis<DumbAction>();
+            chainA[0] = chainB[0] = TestUtils.MineGenesis<DumbAction>(hashAlgorithm);
             for (int i = 1; i < chainA.Length / 2; i++)
             {
                 _branchpoint = chainA[i] = chainB[i] =